Create separate component for navigation items.
